Apply timeout to socat invocations

+++ PrimeCOMAL/solution_2/run.sh
@@ -1,6 +1,8 @@
 export TERM=xterm
 
-socat - EXEC:'opencomalrun opencomal.cml',pty,setsid,ctty < /dev/zero > /dev/null
+timeout -s 9 1m socat - EXEC:'opencomalrun opencomal.cml',pty,setsid,ctty < /dev/zero > /dev/null
 
-cat output.txt
-rm output.txt
+if [ -f output.txt ]; then
+    cat output.txt
+    rm output.txt
+fi
